Use KeyValueForm for `PathParamForm`
