+++ packages/playground/src/components/playground/FormDataForm/index.ts
@@ -1,6 +1,3 @@
 // export { FormDataForm } from "./FormDataForm";
-export {
-  reduceFormDataParameters,
-  enforceTerminalDraftParameter as enforceFormDataTerminalDraftParameter,
-} from "./data";
-export { FormDataParameterSchema } from "./types";
+export { reduceFormDataParameters } from "./data";
+// export { FormDataParameterSchema } from "./types";
