+++ packages/playground/src/components/playground/RequestPanel/PathParamForm/index.ts
@@ -1,1 +1,1 @@
-export { PathParamForm } from "./PathParamForm";
+// export { PathParamForm } from "./PathParamForm";
