+++ packages/playground/src/components/playground/store/request-body.ts
@@ -2,7 +2,7 @@
 // import { FormDataParameterSchema } from "../FormDataForm";
 // import type { KeyValueElement } from "./types";
 
-import type { KeyValueElement } from "./types";
+// import type { KeyValueElement } from "./types";
 
 // const PlaygroundBodyJsonTypeSchema = z.literal("json");
 // const PlaygroundBodyTextTypeSchema = z.literal("text");
